refactor(ptt): extract client app loop
Why:
- Isolate the Client PTT Session loop behind an app::run entrypoint so main.rs is only CLI parsing, tracing, adapter construction, and invocation.
- Bundle daemon, hotkey, Injection, Overlay, and LLM dependencies in ClientPorts so loop behavior can be tested without production devices or sockets.

Side effects:
- main.rs shrinks below the issue threshold; WsClient no longer exposes split helpers because the app loop consumes the higher-level daemon connection port.
