+++ parakeet-ptt/src/client.rs
@@ -1,5 +1,4 @@
 use anyhow::{Context, Result};
-use futures::stream::{SplitSink, SplitStream};
 use futures::{SinkExt, StreamExt};
 use tokio::net::TcpStream;
 use tokio::time::timeout;
@@ -10,9 +9,6 @@ use crate::config::ClientConfig;
 use crate::protocol::{decode_server_message, ClientMessage, DecodedServerMessage, ServerMessage};
 
 pub type WsStream = WebSocketStream<MaybeTlsStream<TcpStream>>;
-pub type WsWrite = SplitSink<WsStream, Message>;
-pub type WsRead = SplitStream<WsStream>;
-
 pub struct WsClient {
     stream: WsStream,
 }
@@ -59,10 +55,6 @@ impl WsClient {
             }
         }
         Ok(None)
-    }
-
-    pub fn into_split(self) -> (WsWrite, WsRead) {
-        self.stream.split()
     }
 
     #[allow(dead_code)]
